+++ contrib/examples/actions/print_to_stdout_and_stderr.sh
@@ -1,3 +1,4 @@
+#!/usr/bin/env bash
 # Licensed to the StackStorm, Inc ('StackStorm') under one or more
 # contributor license agreements.  See the NOTICE file distributed with
 # this work for additional information regarding copyright ownership.
@@ -11,24 +12,21 @@
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
-# limitations under the License.
 
-__all__ = [
-    'DEFAULT_COLLECTION_INTERVAL',
-    'DEFAULT_SLEEP_DELAY',
-    'MINIMUM_TTL_DAYS',
-    'MINIMUM_TTL_DAYS_EXECUTION_OUTPUT'
-]
+COUNT=${1:-"100"}
+SLEEP_DELAY=${2:-"0.5"}
 
+i="0"
+while [ $i -lt ${COUNT} ];
+do
+    j=$[$i+1]
+    if [ $(( $i % 2)) -eq 0 ]; then
+        echo "stderr line ${j}" >&2
+    else
+        echo "stdout line ${j}"
+    fi
 
-# Default garbage collection interval (in seconds)
-DEFAULT_COLLECTION_INTERVAL = 600
+    i=$[$i+1]
 
-# How to long to wait / sleep between collection of different object types (in seconds)
-DEFAULT_SLEEP_DELAY = 2
-
-# Minimum value for the TTL. If user supplies value lower than this, we will throw.
-MINIMUM_TTL_DAYS = 7
-
-# Minimum TTL in days for action execution output objects.
-MINIMUM_TTL_DAYS_EXECUTION_OUTPUT = 1
+    sleep ${SLEEP_DELAY}
+done